Update README.md
Change the rule for participating in the event from "Every task in different language" to "Every day in different language" to ensure participants use a different programming language each day. AI will only be used to suggest a language, not for code generation. Task input and description are versioned in the day folder for documentation purposes.